Create DV inline property from another file.md
Templater template that uses Dataview to display a property from another note.

The user is prompted with with an input box to search for a note. 
When the user has selected a note another input box shows up that will list all the available properties from that specific note. A dataview query will be inserted in the current note.